pyproject.toml
[project]
name = "nuclic-monitoring-system"
version = "0.1.0"
description = "This project is a Telegram bot focused on monitoring the stock of the group's cabinet in the faculty."
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot (>=22.0,<23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
